Fix crash when using a pool template with an entry > 65535, fixes issue #3632
--HG--
branch : trunk

+++ src/server/game/Entities/GameObject/GameObject.cpp
@@ -318,7 +318,7 @@ void GameObject::Update(uint32 diff)
                                 return;
                             }
                                                             // respawn timer
-                            uint16 poolid = GetDBTableGUIDLow() ? sPoolMgr.IsPartOfAPool<GameObject>(GetDBTableGUIDLow()) : 0;
+                            uint32 poolid = GetDBTableGUIDLow() ? sPoolMgr.IsPartOfAPool<GameObject>(GetDBTableGUIDLow()) : 0;
                             if (poolid)
                                 sPoolMgr.UpdatePool<GameObject>(poolid, GetDBTableGUIDLow());
                             else
@@ -560,7 +560,7 @@ void GameObject::Delete()
     SetGoState(GO_STATE_READY);
     SetUInt32Value(GAMEOBJECT_FLAGS, GetGOInfo()->flags);
 
-    uint16 poolid = GetDBTableGUIDLow() ? sPoolMgr.IsPartOfAPool<GameObject>(GetDBTableGUIDLow()) : 0;
+    uint32 poolid = GetDBTableGUIDLow() ? sPoolMgr.IsPartOfAPool<GameObject>(GetDBTableGUIDLow()) : 0;
     if (poolid)
         sPoolMgr.UpdatePool<GameObject>(poolid, GetDBTableGUIDLow());
     else
